+++ src/main/java/ssodam/ssodam/repository/PostRepository.java
@@ -1,6 +1,7 @@
 package ssodam.ssodam.repository;
 
 import ssodam.ssodam.domain.Category;
+import ssodam.ssodam.domain.Member;
 import ssodam.ssodam.domain.Post;
 
 import java.util.List;
