+++ src/main/java/ssodam/ssodam/service/PostService.java
@@ -1,5 +1,6 @@
 package ssodam.ssodam.service;
 
+import ssodam.ssodam.domain.Member;
 import ssodam.ssodam.domain.Post;
 
 import java.util.List;
